Feature: Delete UserDomainService and moved logic to UserApplicationService
